fix file structure add eslint, node ver check
