+++ src/components/Home/Home.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Home = () => {
+  return (
+    <div>
+      home page
+    </div>
+  );
+};
+
+export default Home;
