--- src/pages/index.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import Navbar from '../components/Navbar';
-
-const Home = () => {
-    return (
-        <>
-        <Navbar />
-        </>
-    )
-}
-
-export default Home
